+++ examples/crud.py
@@ -43,8 +43,9 @@ class UserManager(Manager):
 async def main():
     client = RESTClient(httpx.AsyncClient(base_url="https://localhost:8000"))
 
-    response = await UserManager.create(
-        client=client,
+    manager = UserManager(client=client)
+
+    response = await manager.create(
         obj=UserCreateSchema(
             username="admin",
             email="admin@admin.com",
@@ -55,32 +56,31 @@ async def main():
     )
     print(response)  # id=1 username='admin' email='admin@admin.com' age=19
 
-    response = await UserManager.read(
-        client=client,
+    response = await manager.read(
         response_type=UserReadSchema,
     )
 
     for obj in response:
         print(obj)  # id=1 username='admin' email='admin@admin.com' age=19
 
-    response = await UserManager.read_one(
-        client=client,
+    response = await manager.read_one(
         obj_or_pk=1,
         response_type=UserReadSchema,
     )
 
     print(response)  # id=1 username='admin' email='admin@admin.com' age=19
 
-    response = await UserManager.update(
-        client=client,
-        obj=UserUpdateSchema(id=1, username="admin123", ),
+    response = await manager.update(
+        obj=UserUpdateSchema(
+            id=1,
+            username="admin123",
+        ),
         response_type=UserReadSchema,
     )
 
     print(response)  # id=1 username='admin123' email='admin@admin.com' age=19
 
-    await UserManager.delete(
-        client=client,
+    await manager.delete(
         obj_or_pk=1,
         expected_status=204,
     )